fix: compilation errors — notify v7, sysinfo v0.33, tonic v0.12 API changes, collector signature, video type mismatch

+++ src/plugins/sense/video.rs
@@ -20,7 +20,7 @@ pub struct VideoConfig {
 fn default_frame_interval() -> u64 {
     5
 }
-fn default_max_frames() -> u64 {
+fn default_max_frames() -> usize {
     60
 }
 fn default_frame_analyzer() -> String {
